api/requirements.txt
anyio==4.3.0
asgiref==3.8.1
Brotli==1.1.0
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.3
django-cors-headers==4.3.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
mutagen==1.47.0
pycryptodomex==3.20.0
PyJWT==2.8.0
pytz==2024.1
requests==2.31.0
sniffio==1.3.1
sqlparse==0.4.4
tzdata==2024.1
urllib3==2.2.1
websockets==12.0
yt-dlp==2024.3.10
